+++ src/lang/builtins_http.ml
@@ -32,8 +32,15 @@ let add_http_request http name descr request =
   let log = Log.make [name] in
   let header_t = Lang.product_t Lang.string_t Lang.string_t in
   let headers_t = Lang.list_t header_t in
-  let status_t = Lang.tuple_t [Lang.string_t; Lang.int_t; Lang.string_t] in
-  let request_return_t = Lang.tuple_t [status_t; headers_t; Lang.string_t] in
+  let request_return_t =
+    Lang.method_t Lang.string_t
+      [
+        ("protocol_version", ([], Lang.string_t));
+        ("status_code", ([], Lang.int_t));
+        ("status_message", ([], Lang.string_t));
+        ("headers", ([], headers_t));
+      ]
+  in
   let params =
     if List.mem request [Get; Head; Delete] then []
     else [("data", Lang.string_t, Some (Lang.string ""), Some "POST data.")]
@@ -62,7 +69,7 @@ let add_http_request http name descr request =
       in
       let timeout = Lang.to_float (List.assoc "timeout" p) in
       let url = Lang.to_string (List.assoc "" p) in
-      let (x, y, z), headers, data =
+      let (protocol_version, status_code, status_message), headers, data =
         try
           let uri = Http.parse_url url in
           let request =
@@ -86,25 +93,27 @@ let add_http_request http name descr request =
             Printf.sprintf "Error while processing request: %s"
               (Printexc.to_string e) )
       in
-      let status = Lang.tuple [Lang.string x; Lang.int y; Lang.string z] in
+      let protocol_version = Lang.string protocol_version in
+      let status_code = Lang.int status_code in
+      let status_message = Lang.string status_message in
       let headers =
         List.map
           (fun (x, y) -> Lang.product (Lang.string x) (Lang.string y))
           headers
       in
       let headers = Lang.list headers in
-      Lang.tuple [status; headers; Lang.string data])
+      Lang.meth (Lang.string data)
+        [
+          ("protocol_version", protocol_version);
+          ("status_code", status_code);
+          ("status_message", status_message);
+          ("headers", headers);
+        ])
 
 let () =
   let add_http_request = add_http_request (module Http) in
-  add_http_request "http.get"
-    "Perform a full Http GET request and return `(status,headers,data)`." Get;
-  add_http_request "http.post"
-    "Perform a full Http POST request and return `(status,headers,data)`." Post;
-  add_http_request "http.put"
-    "Perform a full Http PUT request and return `(status,headers,data)`." Put;
-  add_http_request "http.head"
-    "Perform a full Http HEAD request and return `(status,headers,data)`." Head;
-  add_http_request "http.delete"
-    "Perform a full Http DELETE request and return `(status,headers,data)`."
-    Delete
+  add_http_request "http.get" "Perform a full Http GET request." Get;
+  add_http_request "http.post" "Perform a full Http POST request`." Post;
+  add_http_request "http.put" "Perform a full Http PUT request." Put;
+  add_http_request "http.head" "Perform a full Http HEAD request." Head;
+  add_http_request "http.delete" "Perform a full Http DELETE request." Delete