+++ src/lang/lang.mli
@@ -27,6 +27,8 @@ val log : Log.t
 (** The type of a value. *)
 type t = Lang_types.t
 
+type scheme = Lang_types.scheme
+
 (** Position in source code. *)
 type pos = Lang_types.pos
 
@@ -230,6 +232,8 @@ val product_t : t -> t -> t
 val of_product_t : t -> t * t
 val tuple_t : t list -> t
 val of_tuple_t : t -> t list
+val record_t : (string * t) list -> t
+val method_t : t -> (string * scheme) list -> t
 val list_t : t -> t
 val of_list_t : t -> t
 val ref_t : t -> t
@@ -278,6 +282,8 @@ val source : Source.source -> value
 val request : Request.t -> value
 val product : value -> value -> value
 val tuple : value list -> value
+val meth : value -> (string * value) list -> value
+val record : (string * value) list -> value
 val reference : value ref -> value
 
 (** Build a function from an OCaml function. Items in the prototype indicate